looking good. Moved main.js from pong to Game.js, created a pages folder to differentiate with the utils. Modules and pages now share the html and init functions principle
