+++ frontend/html/Game.html
@@ -1,8 +1,1 @@
-<div id="game">
-    <canvas id="canvas"></canvas>
-</div>
-    <script type="module" src="/js/pong/main.js"></script>
-    <script type="module">
-		import {displayMainMenu} from "/js/pong/main.js"
-		displayMainMenu();
-    </script>
+<canvas id="canvas"></canvas>
