+++ frontend/html/Sidebar.html
@@ -52,7 +52,7 @@
 			</ul>
 			<a class="d-flex align-items-center text-white text-decoration-none dropdown-toggle" id="dropdownUser1"
 				data-bs-toggle="dropdown" navlink>
-				<img src="/assets/img/leo.jpg" width="30" height="30" class="rounded-circle me-2" />
+				<img width="30" height="30" class="rounded-circle me-2 avatarDynamic" />
 			</a>
 			<span class="section-name ms-2 usernameDynamic" style="position: relative">Boula</span>
 		</div>
