+++ frontend/index.html
@@ -25,7 +25,10 @@
 <body>
 	<div class="d-flex">
 		<div id="sidebar-container"></div>
-		<div id="content-container" class="container-fluid"></div>
+		<div id="content-container" class="container-fluid">
+			<div id="content"></div>
+			<div id="game" hidden></div>
+		</div>
 	</div>
 </body>
 </html>
