+++ frontend/js/pong/online.js
@@ -6,7 +6,7 @@ import { initGame } from "./initGame.js";
 import { translateBall} from "./onlineCollision.js";
 import { handlerScore, setBallData, handlerStatusMessage } from "./handlerMessage.js";
 import { sendCharacter} from "./sendMessage.js";
-import { characters } from "./main.js";
+import { characters } from "../pages/Game.js";
 import { updateMixers } from "./displayCharacter.js";
 import { resize } from "./resize.js";
 import * as THREE from 'three';
